+++ src/geff/graph_libs/_networkx.py
@@ -6,12 +6,12 @@ import networkx as nx
 
 from geff.core_io import write_dicts
 from geff.core_io._base_read import read_to_memory
-from geff.io_utils import (
-    calculate_roi_from_nodes,
+from geff.core_io.utils import calculate_roi_from_nodes
+from geff.metadata.schema import GeffMetadata, _axes_from_lists
+from geff.metadata.utils import (
     create_or_update_metadata,
     get_graph_existing_metadata,
 )
-from geff.metadata.schema import GeffMetadata, _axes_from_lists
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
